fix: 修改库存记录表的数量字段类型
1. 将 TileStockRecord 实体类中的数量字段类型从 Integer 改为 Long
2. 修改数据库表结构，将 quantity 字段类型改为 bigint

+++ ruoyi-system/src/main/java/com/ruoyi/system/tile/domain/TileStockRecord.java
@@ -37,7 +37,7 @@ public class TileStockRecord extends BaseEntity
 
     /** 数量 */
     @Excel(name = "数量")
-    private Integer quantity;
+    private Long quantity;
 
     /** 来源单据ID */
     private Long sourceId;
@@ -118,12 +118,12 @@ public class TileStockRecord extends BaseEntity
         return warehouseName;
     }
 
-    public void setQuantity(Integer quantity) 
+    public void setQuantity(Long quantity) 
     {
         this.quantity = quantity;
     }
 
-    public Integer getQuantity() 
+    public Long getQuantity() 
     {
         return quantity;
     }
